Changed amount type to decimal

+++ app/Actions/Activity/AddActivityAction.php
@@ -12,7 +12,7 @@ class AddActivityAction
     ) {}
 
     public function __invoke(
-        int $amount,
+        float $amount,
         ?string $note,
         int $categoryId
     ): Activity
